Moved getStoragePercent into OS specific library files
